backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Pillow==10.4.0
google-genai==0.1.2
python-dotenv==1.0.1
gunicorn==22.0.0
